Issues resolved related to ver.2.2.2 bug list. Added Parental Control to Help & Info.

+++ Moviebeam/app/src/main/java/com/diipl/moviebeam/ui/kappingservice/EndlessService.kt
@@ -331,6 +331,10 @@ class EndlessService : Service() {
                         }
                     }
 
+                    Intent.ACTION_MEDIA_BUTTON -> {
+                        Log.e(TAG, "onReceive: ACTION_MEDIA_BUTTON")
+                    }
+
                     else -> {
                         Log.e(TAG, "onReceive: ${it.action}")
                     }
@@ -425,7 +429,7 @@ class EndlessService : Service() {
                         pingFakeServer()
                         callKapingApi()
 
-                        if (Constants.SESSION_ID.isNotEmpty())
+                        if (Constants.SESSION_ID.isNotEmpty() && Constants.SESSION_ID != "null")
                             roomRepository.removeOverTimeMovies()
 
                         if (Constants.SESSION_ID == "null") {
